requirements.txt
Flask==3.1.2
kafka_python==2.2.15
langchain_core==0.3.74
langchain_mistralai==0.2.11
python-dotenv==1.1.1
